feat: Order, Comment 타입 types.ts에서 정의
