+++ types.ts
@@ -1,0 +1,11 @@
+export interface Comment {
+  id: number;
+  content: string;
+  updatedAt: number;
+  writer: {
+    image: string;
+    nickname: string;
+  };
+}
+
+export type Order = "recent" | "like";
